Feat: 전반적으로 DTO, Converter, Exception, ErrorCode, Service Command/Query 분리 적용 및 Test 도메인 추가 - Member, Mission, Review, Store 도메인에 대해 DTO를 RequestDTO와 ResponseDTO로 분리 - Converter 클래스 추가하여 Entity와 DTO 간의 변환 로직 분리 - Exception 클래스와 ErrorCode 클래스 추가하여 예외 처리 구조 개선 - Service 레이어에서 Command와 Query로 역할 분리 - Test 도메인 추가하여 기본적인 CRUD 기능 구현 및 테스트 - 글로벌 API 응답 및 예외 처리 구조 추가 - .idea/gradle.xml 및 build.gradle 파일 수정 - 기존 DTO 클래스 파일명 변경 및 불필요한 DTO 파일 삭제 - 관련 컨트롤러, 서비스, 리포지토리 클래스 수정

+++ src/main/java/spring/umc/domain/review/repository/ReviewRepositoryImpl.java
@@ -24,7 +24,7 @@ public class ReviewRepositoryImpl implements ReviewRepositoryCustom {
     private final JPAQueryFactory queryFactory;
 
     @Override
-    public Page<Review> searchReviews(Long memberId, Long storeId, Integer minScore, Pageable pageable) {
+    public Page<Review> searchReviews(Long memberId, Long storeId, Double minScore, Pageable pageable) {
 
         BooleanBuilder builder = new BooleanBuilder();
